Move confighandle package contents to config.handle.
