+++ src/main/java/org/mvplugins/multiverse/core/command/MVCommandCompletions.java
@@ -32,7 +32,7 @@ import org.jvnet.hk2.annotations.Service;
 import org.mvplugins.multiverse.core.anchor.AnchorManager;
 import org.mvplugins.multiverse.core.config.CoreConfig;
 import org.mvplugins.multiverse.core.confighandle.functions.DefaultSuggesterProvider;
-import org.mvplugins.multiverse.core.confighandle.PropertyModifyAction;
+import org.mvplugins.multiverse.core.config.handle.PropertyModifyAction;
 import org.mvplugins.multiverse.core.destination.DestinationInstance;
 import org.mvplugins.multiverse.core.destination.DestinationsProvider;
 import org.mvplugins.multiverse.core.destination.core.WorldDestination;
