+++ src/main/java/org/mvplugins/multiverse/core/config/CoreConfig.java
@@ -17,8 +17,8 @@ import org.jvnet.hk2.annotations.Service;
 import org.mvplugins.multiverse.core.MultiverseCore;
 import org.mvplugins.multiverse.core.command.MVCommandManager;
 import org.mvplugins.multiverse.core.command.queue.ConfirmMode;
-import org.mvplugins.multiverse.core.confighandle.CommentedConfigurationHandle;
-import org.mvplugins.multiverse.core.confighandle.StringPropertyHandle;
+import org.mvplugins.multiverse.core.config.handle.CommentedConfigurationHandle;
+import org.mvplugins.multiverse.core.config.handle.StringPropertyHandle;
 import org.mvplugins.multiverse.core.confighandle.migration.BooleanMigratorAction;
 import org.mvplugins.multiverse.core.confighandle.migration.ConfigMigrator;
 import org.mvplugins.multiverse.core.confighandle.migration.IntegerMigratorAction;
