+++ src/main/java/org/mvplugins/multiverse/core/config/handle/BaseConfigurationHandle.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.confighandle;
+package org.mvplugins.multiverse.core.config.handle;
 
 import java.util.HashMap;
 import java.util.List;
