+++ src/main/java/org/mvplugins/multiverse/core/config/handle/MemoryConfigurationHandle.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.confighandle;
+package org.mvplugins.multiverse.core.config.handle;
 
 import io.vavr.control.Try;
 import org.bukkit.configuration.ConfigurationSection;
