+++ src/main/java/org/mvplugins/multiverse/core/config/handle/YamlConfigurationHandle.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.confighandle;
+package org.mvplugins.multiverse.core.config.handle;
 
 import java.io.IOException;
 import java.nio.file.Path;
