+++ src/main/java/org/mvplugins/multiverse/core/world/helpers/DataStore.java
@@ -8,7 +8,7 @@ import io.vavr.control.Try;
 import org.bukkit.GameRule;
 import org.bukkit.World;
 import org.jvnet.hk2.annotations.Service;
-import org.mvplugins.multiverse.core.confighandle.StringPropertyHandle;
+import org.mvplugins.multiverse.core.config.handle.StringPropertyHandle;
 import org.mvplugins.multiverse.core.world.LoadedMultiverseWorld;
 
 /**
